SD Cache: moved to diskio layer, invalidation in case of error
